+++ CHANGELOG.md
@@ -7,9 +7,6 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ## [Unreleased](https://gitlab.gistools.geog.uni-heidelberg.de/climate-action/plugins/ghg-budget/-/compare/dummy...main?from_project_id=854&straight=false)
 ### Changed
-This also means that a postgres server, that serves as the result backend for Celery, is now required to run the
-plugin.
-- use renamed BaseOperator instead of the old Operator class to inherit custom operators from
 - account for new explicit AOI input by moving it from the input parameters to the compute method input
 - simplify the plugin module with the new functionality in climatoology and now using the celery library
 - Improved documentation and descriptions ([#14](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/14), [#17](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/17))
@@ -23,6 +20,7 @@ plugin.
 - Updated emission data for Heidelberg ([#5](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/5))
 - Simple version of GHG budget plugin ([#7](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/7))
 - Chart showing cumulative emissions of Heidelberg ([#27](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/27))
+- Chart showing different possible emission reduction paths for Heidelberg ([#28](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/28))
 - Geoblocker for Heidelberg ([#33](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/33))
 - Add commit hash to plugin version in dockerfile ([#37](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/37))
 - Add test coverage checks to CI pipeline ([#36](https://gitlab.heigit.org/climate-action/plugins/ghg-budget/-/issues/36))